+++ scratch.js
@@ -1,6 +1,6 @@
 'use strict';
 var nlp = require('./src/index');
-nlp.verbose('tagger');
+// nlp.verbose('tagger');
 // const corpus = require('nlp-corpus');
 // let text = corpus.sotu.parsed()[0];
 // const fresh = require('./test/unit/lib/freshPrince.js');
@@ -9,5 +9,10 @@ nlp.verbose('tagger');
 // .out('matches')
 // .out('list')
 
-var doc = nlp('he ate 100 dollars');
-doc.values().debug();
+// var r = nlp('be integrated?');
+// console.log(r.verbs());
+
+var doc = nlp('he is fun');
+console.log('==================');
+doc.match('sdf?');
+console.log('.done.');